Added comments and moved initialization code for mechanics benchmarks
I moved the initialization code for `KanesMethod` and `LagrangesMethod`
to the setup code outside of the tests at Jason's suggestion. Now the
test will exclusively depend on the formation of the equations of
motion. This change allowed much fewer variables to be attached to the
class. I also added commenting to the `KanesMethodSuite` code.